% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSAlgo.R
\name{BuildRegressionEstimateTable}
\alias{BuildRegressionEstimateTable}
\title{Run DHS analysis at a top level}
\usage{
BuildRegressionEstimateTable(
  DataMod,
  RunOn,
  SplitOn = "WWTP",
  verbose = FALSE,
  PSigTest = TRUE
)
}
\arguments{
\item{DataMod}{The DF containing the col RunOn + date}

\item{RunOn}{The col names of the values we wish to run}

\item{SplitOn}{A category to separate to create independent TS data}

\item{verbose}{Bool on whether it should print out what group it is on}

\item{PSigTest}{When categorizing if it should reject high pVals}
}
\value{
A DF with the associated Date and DHS analysis
}
\description{
Run DHS analysis at a top level
}
