% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildWorkSheet4.R
\name{BuildWorkSheet4}
\alias{BuildWorkSheet4}
\title{Convert DHS data to workset4 shape}
\usage{
BuildWorkSheet4(df)
}
\arguments{
\item{df}{DF stored in vignette data}
}
\value{
data frame
}
\description{
Convert DHS data to workset4 shape
}
