% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierTools.R
\name{computeJumps}
\alias{computeJumps}
\title{computeJumps}
\usage{
computeJumps(df)
}
\arguments{
\item{n/a}{}
}
\value{
dataframe with 4 columns appended: delta(n1), delta(n2) from left and right
}
\description{
computeJumps
}
