% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{ConfMatrix}
\alias{ConfMatrix}
\title{ConfMatrix}
\usage{
ConfMatrix(DF, Cat, x, y)
}
\arguments{
\item{DF}{data frame containing results of DHS analysis}

\item{Cat}{The column with the values of the methods}

\item{x}{The first method to compare}

\item{y}{The second method to compare}
}
\value{
a ggplot object of the confusion matrix
}
\description{
creates a confusion matrix from data long format
}
