% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{createDHSMethod_Plot}
\alias{createDHSMethod_Plot}
\title{Create representative plot of the DHS analysis}
\usage{
createDHSMethod_Plot(
  RegDF,
  BaseDF,
  FacGridFormula = Method ~ WWTP,
  SiteName = "WWTP",
  PointVal = "sars_cov2_adj_load_log10",
  LineVal = NULL
)
}
\arguments{
\item{RegDF}{DF containing the regression analysis}

\item{BaseDF}{The DF containing the raw data}

\item{FacGridFormula}{The formula we wish to facet the heatmaps with}

\item{SiteName}{The column names for Site}

\item{PointVal}{The point columns we want to plot}

\item{LineVal}{The Line columns we want to plot}
}
\value{
a ggplot of the heatmap of each method and the underlying data
}
\description{
Create representative plot of the DHS analysis
}
\examples{

data(example_data, package = "DSIWasteWater")
example_reg_table <- buildRegressionEstimateTable(example_data)
createDHSMethod_Plot(example_reg_table, example_data)
}
