% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSAlgo.R
\name{CreateHeatMaps}
\alias{CreateHeatMaps}
\title{CreateHeatMaps}
\usage{
CreateHeatMaps(DF, ToMerge = FALSE)
}
\arguments{
\item{DF}{}

\item{ToMerge}{if true we remove the lower labels}
}
\value{
faceted ggplot
}
\description{
Creates graphic of model prediction for each method
}
