% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{CreateHeatMaps_Plot}
\alias{CreateHeatMaps_Plot}
\title{CreateHeatMaps_Plot}
\usage{
CreateHeatMaps_Plot(
  DF,
  FacGridFormula,
  FillFac,
  CatagoryColors,
  ToMerge = FALSE
)
}
\arguments{
\item{DF}{The DF used to create the Heatmap}

\item{FacGridFormula}{how we wan to facet the heatmap}

\item{FillFac}{the name of the catagory method}

\item{CatagoryColors}{The color scheme used}

\item{ToMerge}{if true we remove the lower labels}
}
\value{
faceted ggplot heatmap
}
\description{
Creates graphic of model prediction for each method
}
