% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_wastewater.R
\docType{data}
\name{Data_wastewater}
\alias{Data_wastewater}
\title{WasteWater data set cleaned from the surveillance branch}
\format{
Rows: 8373 Columns: 30
}
\usage{
data("Data_wastewater")
}
\description{
Data from the DHS of waste water prevalence in Wisconsin cites
}
\keyword{datasets}
