% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSAlgo.R
\name{DHSClassificationFunc}
\alias{DHSClassificationFunc}
\title{DHSClassificationFunc}
\usage{
DHSClassificationFunc(DF, PSigTest = TRUE)
}
\arguments{
\item{DF}{}

\item{PSigTest}{Controls if we filter values with P-values>.3}
}
\description{
Adds the DHS Classification scheme to data created by DHSOuterLoop
}
