% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSAlgo.R
\name{DHSInnerLoop}
\alias{DHSInnerLoop}
\title{DHSInnerLoop}
\usage{
DHSInnerLoop(Formula, DF, Keep = NULL, LMMethod = lm)
}
\arguments{
\item{Formula}{LM model to be fit on DF}

\item{DF}{Contains the data needed for Formula to work}

\item{LMMethod}{Controls what Linear model is applied. 
intended options are lm and FCVLM}
}
\value{
a row of a DF containing the 
WWTP, last date, timespan, number of rows, model slope and significance,
and predicted percent change, and what linear model was used
}
\description{
Runs a Linear regression on the data and returns it in a form to be merged
in the DHSOuterLoop function
}
