% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSTrendGen.R
\name{ExpSmoothMod}
\alias{ExpSmoothMod}
\title{ExpSmoothMod
Add a column of the smoothed values using exponential smoothing}
\usage{
ExpSmoothMod(DF, InVar, OutVar, alpha = "guess", beta = "guess", Filter = NULL)
}
\arguments{
\item{DF}{The DF we are to add a exponential smoothing column to}

\item{InVar}{The column to be smoothed}

\item{OutVar}{The name of the new column}

\item{alpha}{The alpha fed into robets exponential smoothing. if it equals "guess" then it
if found using ParameterGuess}

\item{beta}{The beta fed into robets exponential smoothing. if it equals "guess" then it
if found using ParameterGuess}

\item{Filter}{Prefilter using the value of a Filter col}
}
\value{
A DF with an extra col with a exp smoothed version of InVar
}
\description{
ExpSmoothMod
Add a column of the smoothed values using exponential smoothing
}
