% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMVariants.R
\name{FCVLM}
\alias{FCVLM}
\title{FCVLM}
\usage{
FCVLM(Formula, data)
}
\arguments{
\item{Formula}{LM model to be fit on robust data}

\item{data}{Contains the data needed for Formula to work}
}
\value{
the best of the lm models by p values
}
\description{
fit n-1 models droping 1 data point
returns the model with the lowest p-val form among the,
}
