% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMVariants.R
\name{LMSelectBestList}
\alias{LMSelectBestList}
\title{LMSelectBestList}
\usage{
LMSelectBestList(LMList, Verbose = FALSE)
}
\arguments{
\item{LMList}{List of possible linear models}

\item{Verbose}{prints the row removed and the p value}
}
\value{
the LM with the lowest P-value
}
\description{
LMSelectBestList
}
