% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSTrendGen.R
\name{NGuess}
\alias{NGuess}
\title{NGuess for sgolaySmoothMod number of points per polynomial}
\usage{
NGuess(DF, InVar, Base, min)
}
\arguments{
\item{DF}{The DF to fit the sgolayfilt curve on. should only
be from one TS}

\item{InVar}{The col name that the variable is about to smooth}

\item{Base}{The value we are scaling by the of the number of
rows in the data}

\item{min}{The smallest value returned}
}
\value{
a number greater or equal to min
}
\description{
NGuess for sgolaySmoothMod number of points per polynomial
}
\examples{
NGuess(DF,InVar,50/178, 7)
}
