% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierTools.R
\name{rankJumps}
\alias{rankJumps}
\title{rankJumps}
\usage{
rankJumps(df)
}
\arguments{
\item{df}{DataFrame. needs Column n1.jumpFromLeft, n1.jumpFromRight, 
          n2.jumpFromLeft, n2.jumpFromRight, WWTP
          
First 4 gen from computeJumps}
}
\value{
dataframe with 4 columns appended: ranks of each of the 4 jumps;
}
\description{
Convert jumps from last step into a ordering
}
