% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierDetectionFuncs.R
\name{TrendSDOutlierDetec}
\alias{TrendSDOutlierDetec}
\title{Decide outliers using calculated trend and standard deviation}
\usage{
TrendSDOutlierDetec(
  DF,
  VecName,
  SDDeg,
  DaySmoothed = 36,
  n = 5,
  TrendFunc = LoessSmoothMod
)
}
\arguments{
\item{DF}{The DF Needs to have the Time label to be Date Needs to be sorted by 
Date before called}

\item{VecName}{String name of the variable we are looking at outliers for}

\item{SDDeg}{How many Standard deviations away from the trend to flag}

\item{DaySmoothed}{Days used to get Standard deviation}

\item{n}{Number of iterations of method done}

\item{TrendFunc}{What function that is used to generate the trend}
}
\value{
Boolean of each row being an outlier
}
\description{
Need to be Done
}
