% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierDetectionFuncs.R
\name{TrendSDOutlierFilter}
\alias{TrendSDOutlierFilter}
\title{wrapper for TrendSDOutlierDetec that stores outlier results}
\usage{
TrendSDOutlierFilter(
  DF,
  VecName,
  SDDeg,
  DaySmoothed,
  outCol = "FlaggedOutlier",
  n = 5,
  TrendFunc = LoessSmoothMod,
  verbose = FALSE
)
}
\arguments{
\item{DF}{The DF Needs to have the Time label to be Date}

\item{VecName}{String name of the variable we are looking at outliers for}

\item{SDDeg}{How many Standard deviations away from the trend to flag}

\item{DaySmoothed}{Days used to get Standard deviation}

\item{outCol}{What the name of the outVec is}

\item{n}{Number of iterations of method done}

\item{TrendFunc}{What function that is used to generate the trend}

\item{verbose}{Controls if it prints the \% of data that is outliers}
}
\value{
DF With col saying whether the method was flagged
}
\description{
in the data frame. The input assumes the data is from one TS
}
