% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{WastePlot}
\alias{WastePlot}
\title{Title}
\usage{
WastePlot(
  DF,
  xVal,
  PointVal,
  LineVal,
  PointName = NULL,
  LineName = NULL,
  ToMerge = FALSE
)
}
\arguments{
\item{DF}{}

\item{xVal}{}

\item{PointVal}{}

\item{LineVal}{}

\item{PointName}{}

\item{LineName}{}

\item{ToMerge}{}
}
\description{
Title
}
