% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{Abstract_PlotAdd}
\alias{Abstract_PlotAdd}
\title{Title}
\usage{
Abstract_PlotAdd(GGObj, GGfunc, YName, YVal)
}
\arguments{
\item{GGObj}{}

\item{GGfunc}{}

\item{YName}{}

\item{YVal}{}
}
\description{
Title
}
