% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{createDHSMethod_Plot}
\alias{createDHSMethod_Plot}
\title{Create representative plot of the DHS analysis}
\usage{
createDHSMethod_Plot(
  RegDF,
  BaseDF,
  FacGridFormula = Method ~ WWTP,
  SiteName = "WWTP",
  PointVal = "sars_cov2_adj_load_log10",
  LineVal = NULL,
  ncol = 3
)
}
\arguments{
\item{RegDF}{data frame containing the regression analysis}

\item{BaseDF}{The data frame containing the raw data}

\item{FacGridFormula}{The formula we wish to facet the heat maps with}

\item{SiteName}{The column names for Site}

\item{PointVal}{The point columns we want to plot}

\item{LineVal}{The Line columns we want to plot}

\item{ncol}{The number of plots in each row}
}
\value{
a ggplot of the heat map of each method and the underlying data
}
\description{
createDHSMethod_Plot uses RegDF to create the top set of plots showing the 
predictions at each time period. 
It uses BaseDF to create the lower plot that shows what the regressed data
looks like
}
\examples{

data(example_data, package = "DSIWasteWater")
example_reg_table <- buildRegressionEstimateTable(example_data)
createDHSMethod_Plot(example_reg_table, example_data)
}
