% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{createWasteGraph_Plot}
\alias{createWasteGraph_Plot}
\title{Waste Water graphic}
\usage{
createWasteGraph_Plot(
  DF,
  xVal,
  PointVal = NULL,
  LineVal = NULL,
  ToMerge = FALSE
)
}
\arguments{
\item{DF}{DF containing waste water measurements specified in the remaining params}

\item{xVal}{name of x variable, normally close to "Date"}

\item{PointVal}{the discrete measurements}

\item{LineVal}{the continuous measurements}

\item{ToMerge}{remove facet info if true. be careful that the to plots have same ordering}
}
\value{
a ggplot object with points with lables for each PointVal and a lines for each LineVal
}
\description{
Waste Water graphic
}
