% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{DHSTopLevelPlots}
\alias{DHSTopLevelPlots}
\title{Title}
\usage{
DHSTopLevelPlots(
  RegDF,
  BaseDF,
  FacGridFormula = Method ~ WWTP,
  SiteName = "WWTP",
  PointName = "sars_cov2_adj_load_log10",
  LineName = NULL
)
}
\arguments{
\item{RegDF}{}

\item{BaseDF}{}

\item{FacGridFormula}{}

\item{SiteName}{}

\item{PointName}{}

\item{LineName}{}
}
\description{
Title
}
