% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{DHSTopLevelPlots}
\alias{DHSTopLevelPlots}
\title{Title}
\usage{
DHSTopLevelPlots(RegDF, BaseDF, FacGridFormula, SiteName, PointName, LineName)
}
\arguments{
\item{RegDF}{}

\item{BaseDF}{}

\item{FacGridFormula}{}

\item{SiteName}{}

\item{PointName}{}

\item{LineName}{}
}
\description{
Title
}
