% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSAlgo.R
\name{LMDropList}
\alias{LMDropList}
\title{LMDropList}
\usage{
LMDropList(n, Formula, data)
}
\arguments{
\item{n}{the index of the row to drow}

\item{Formula}{LM model to be fit on robust data}

\item{data}{Contains the data needed for Formula to work}
}
\description{
fits a lm model with the nth row of data removed
}
