% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSTrendGen.R
\name{LoessSmoothMod}
\alias{LoessSmoothMod}
\title{LoessSmoothMod
Add a column of the smoothed values using Loess}
\usage{
LoessSmoothMod(DF, InVar, OutVar, span = "guess", Filter = NULL)
}
\arguments{
\item{DF}{}

\item{InVar}{The column to be smoothed}

\item{OutVar}{The name of the new column}

\item{span}{The span fed into loess smoothing. if it equals "guess" then it
if found using ParameterGuess}

\item{Filter}{Prefilter using the value of a Filter col}
}
\value{
A DF with an extra col with a loesss smoothed version of InVar
}
\description{
LoessSmoothMod
Add a column of the smoothed values using Loess
}
