% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{orderAndZipListsOfPlots_Plot}
\alias{orderAndZipListsOfPlots_Plot}
\title{Take two list of plots and combine}
\usage{
orderAndZipListsOfPlots_Plot(top_plot_list, bot_plot_list, ratA = 3, ratB = 1)
}
\arguments{
\item{top_plot_list}{}

\item{bot_plot_list}{}

\item{ratA}{}

\item{ratB}{}
}
\description{
Take two list of plots and combine
}
