% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{WastePlot}
\alias{WastePlot}
\title{Waste Water graphic}
\usage{
WastePlot(DF, xVal, PointValVec = NULL, LineValVec = NULL, ToMerge = FALSE)
}
\arguments{
\item{DF}{DF containing waste water measurements specified in the remaining params}

\item{xVal}{name of x variable, normally close to "Date"}

\item{PointValVec}{the discrete measurements}

\item{LineValVec}{the continuous measurements}

\item{ToMerge}{remove facet info if true. be careful that the to plots have same ordering}
}
\value{
a ggplot object with points with lables for each PointValVec and a lines for each LineValVec
}
\description{
Waste Water graphic
}
