% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSPlots.R
\name{WastePlot}
\alias{WastePlot}
\title{Waste Water graphic}
\usage{
WastePlot(
  DF,
  xVal,
  PointVal,
  LineVal,
  PointName = NULL,
  LineName = NULL,
  ToMerge = FALSE
)
}
\arguments{
\item{DF}{DF containing waste water measurements specified in the remaining params}

\item{xVal}{name of x variable, normally close to "Date"}

\item{PointVal}{the discrete measurement}

\item{LineVal}{the continuous measurement}

\item{PointName}{the label of PointVal}

\item{LineName}{the label of LineVal}

\item{ToMerge}{remove facet info if true. be careful that the to plots have same ordering}
}
\description{
Waste Water graphic
}
